start refactoring

+++ main_app/src/app/presentation/web_api/api/event.py
@@ -1,7 +1,7 @@
 from pprint import pprint
 
+from app.adapters.s3.s3_client import S3Client
 from app.adapters.ws.ws_client import webscocket_client
-from app.s3.s3 import S3Client
 from fastapi import APIRouter, Request
 
 s3_storage_out = S3Client(
@@ -20,10 +20,13 @@ async def get_event(request: Request):
     data = await request.json()
     filename = data.get("Records")[0].get("s3").get("object").get("key")
     ws = webscocket_client.get_file(filename)
+    if not ws:
+        return
 
     url = await s3_storage_out.get_url_for_file(filename)
     await ws.send_text(url)
     webscocket_client.remove_file(filename)
+    ws.close()
     print(url)
 
     # pprint(data.get("Records")[0].get("s3").get("object").get("key"))
